+++ src/pages/FallBack.tsx
@@ -1,0 +1,5 @@
+import Loading from "elements/Loading";
+
+const FallBack = () => <Loading isLoading />;
+
+export default FallBack;
