--- src/pages/Fallback/index.tsx
@@ -1,10 +1,0 @@
-import Loading from "elements/Loading";
-import { Wrapper } from "./style";
-
-const Fallback = () => (
-  <Wrapper>
-    <Loading isLoading />
-  </Wrapper>
-);
-
-export default Fallback;
