--- src/pages/Fallback/style.ts
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-export const Wrapper = styled.div`
-  min-height: 100vh;
-`;
