+++ src/router/Routers.tsx
@@ -8,8 +8,8 @@ const SignUp = lazy(() => import("pages/SignUp"));
 const SignIn = lazy(() => import("pages/SignIn"));
 const My = lazy(() => import("pages/My"));
 const Find = lazy(() => import("pages/Find"));
-const UsedBook = lazy(() => import("pages/UsedBook"));
 const Oauth = lazy(() => import("pages/Oauth"));
+const UsedBook = lazy(() => import("pages/UsedBook"));
 const Book = lazy(() => import("pages/Book"));
 const Order = lazy(() => import("pages/Order"));
 const Community = lazy(() => import("pages/Community"));
